moved types to their component modules
