+++ src/front/src/components/Container/Container.tsx
@@ -2,8 +2,6 @@ import React from 'react';
 
 import './Container.scss';
 
-const Container: React.FC = ({children}) => {
+export const Container: React.FC = ({children}) => {
   return <div className="container">{children}</div>;
 };
-
-export default Container;
