+++ src/front/src/stores/index.ts
@@ -3,7 +3,11 @@ import React from 'react';
 
 import {default as LinksStore} from './Links';
 import {default as UserStore} from './User';
-import {IStores} from 'types';
+
+export interface IStores {
+  userStore: UserStore;
+  linksStore: LinksStore;
+}
 
 export const initialStores = {
   linksStore: new LinksStore(),
